+++ dev/go.mod
@@ -1,3 +1,18 @@
 module github.com/xataio/pgroll/dev
 
 go 1.23.3
+
+require (
+	github.com/go-echarts/go-echarts/v2 v2.4.5
+	github.com/spf13/cobra v1.8.1
+	github.com/stretchr/testify v1.9.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
